+++ project/backend/bike-rent-potepa-patla/src/main/java/com/bike_rent_potepa_patla/service/CustomerService.java
@@ -2,7 +2,6 @@ package com.bike_rent_potepa_patla.service;
 
 import com.bike_rent_potepa_patla.dto.customer.CustomerCreateDto;
 import com.bike_rent_potepa_patla.dto.customer.CustomerResponseDto;
-import com.bike_rent_potepa_patla.model.Customer;
 import com.bike_rent_potepa_patla.repository.CustomerRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -15,7 +14,7 @@ public class CustomerService {
 
     @Transactional
     public CustomerResponseDto addCustomer(CustomerCreateDto dto) {
-        Long newCustomerId=customerRepository.addNewCustomer(dto.firstName(),dto.surrName(),dto.phone());
+        Integer newCustomerId=customerRepository.addNewCustomer(dto.firstName(),dto.surrName(),dto.phone());
         return CustomerResponseDto.builder()
                 .customerId(newCustomerId)
                 .firstName(dto.firstName())
